+++ src/ts/util/Options.ts
@@ -44,6 +44,7 @@ export class Options {
                 autoSpace: true,
                 chinesePunct: true,
                 fixTermTypo: true,
+                toc: true,
             },
             math: {
                 engine: "KaTeX",