+++ CHANGELOG.md
@@ -44,6 +44,10 @@
 
 ### v2.2.19 / 未发布
 
+* [121](https://github.com/Vanessa219/vditor/issues/121) 所见即所得模式支持 [ToC] `引入特性`
+* [119](https://github.com/Vanessa219/vditor/issues/119) 所见即所得模式支持脚注 `引入特性`
+* [55](https://github.com/Vanessa219/vditor/issues/55) 所见即所得模式支持链接引用定义`引入特性`
+
 ### v2.2.18 / 2020-03-05
 
 * [209](https://github.com/Vanessa219/vditor/pull/209) 支持 Graphviz `引入特性`
@@ -121,6 +125,7 @@
   * 引用快捷键修改
   * 移除 index-preview.html, index-preview.js 文件
   * 添加 graphvizRender
+  * 添加 option.preview.markdown.toc
 
 ### v2.1.15 / 2020-02-09